ca commence a marcher bien
git-svn-id: file:///Users/ng/svn-local/to_share/eftp@6 5591a063-198b-498d-bffb-dae67d11ee29

+++ server.c
@@ -11,6 +11,7 @@
 #include <time.h>
 #include <signal.h>
 #include <sys/wait.h>
+#include <errno.h>
 
 #include "assert.h"
 #include "vector.h"
@@ -76,13 +77,16 @@ int start_server(user_pool_t * eu, port_t port)
     {
 
 	socklen_t addrlen = sizeof(rmaddr);
-	    
+	printf("avt\n");
 	int fd = accept(fdlisten,
 			(struct sockaddr *)&rmaddr,
 			&addrlen);
-	    
+	printf("apr\n");
 	if(fd == -1)
-	    perror("accept");
+	{
+	    if(errno != EINTR)
+		perror("accept");
+	}
 	else
 	{
 
@@ -92,17 +96,21 @@ int start_server(user_pool_t * eu, port_t port)
 	    pid_t c = fork();
 
 	    if(c == 0)
+	    {
 	     /* fils */
+		close(fdlisten);
 		process_new_child(fd);
+		dbg_printf("child halted\n");
+		return EXIT_SUCCESS;
+	    }
 	    else if(c == -1)
 		c_warning2(0, "fork error");
 	    else
 	    { /* pere */
+		close(fd);
 	    }
 	    
 	}
-	    
-	
 
     }
 
@@ -116,6 +124,10 @@ int start_server(user_pool_t * eu, port_t port)
 
 void stop_server()
 {
+    static int count = 0;
+    count++;
+    if(count > 1) return;
+
     server_run = 0;
     kill(0, SIGTERM);
     while(wait(NULL) != -1);
@@ -135,11 +147,12 @@ int process_new_child(int fd)
     cmd_t cmd;
     cmd.user = &user;
 
-    if((ret = send_answer(fd, A_OK, "waiting for user login")) < 0)
+    if((ret = send_answer(fd, A_OK, 0, "waiting for user login and password")) < 0)
        return ret;
 
-
     while((ret = get_answer(&cmd)) > -1);
+
+    close(fd);
 
     return ret;
 }
@@ -155,13 +168,38 @@ int get_answer(cmd_t * cmd)
     if( (recvallline(cmd->user->fd, buff, BUFFS)) < -1)
 	return -1;
 
+    dbg_printf("rec command=%s\n", buff);
+
     command_from_string(buff, cmd);
 
-    return execute_command(cmd);
+    dbg_printf("cmd type=%d\n", cmd->type);
+
+    if(cmd->type == C_ERROR)
+	return send_answer(cmd->user->fd, A_ERROR, RC_BAD_CMD, "bad command or bad parameters");
+    else
+    {
+	switch(execute_command(cmd))
+	{
+	case RC_ACCESS_DENIED:
+	    send_answer(cmd->user->fd, A_ERROR, RC_ACCESS_DENIED, "access denied");
+	    return 0;
+	case RC_NO_AUTH:
+	    send_answer(cmd->user->fd, A_ERROR, RC_NO_AUTH, "not logged in");
+	    return 0;
+	case RC_CMD_ERR:
+	    send_answer(cmd->user->fd, A_ERROR, RC_CMD_ERR, "command error");
+	    return 0;
+	case RC_OK:
+	    return 0;
+	case RC_QUIT:
+	default:
+	    return -1;
+	}
+    }
 }
 
 
-int send_answer(int fd, ans_t a, char * txt)
+int send_answer(int fd, ans_t a, char code, char * txt)
 {
     enum { BUFFS = 12 + MAX_MSG_LEN };
 
@@ -169,8 +207,8 @@ int send_answer(int fd, ans_t a, char * txt)
 
     c_assert(txt == NULL || strlen(txt) < MAX_MSG_LEN);
 
-    snprintf(buff, BUFFS, "%d %s %s\n",
-	     a,
+    snprintf(buff, BUFFS, "%d%d %s %s\n",
+	     a, code,
 	     a == A_ERROR ? "error" : "ok",
 	     txt ? txt : ""
 	    );
